backend/requirements1.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.8.0
astunparse==1.6.3
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.17.0
firebase-admin==6.6.0
flatbuffers==25.1.21
fsspec==2025.3.0
gast==0.6.0
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-genai==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
mypy==1.14.1
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
openai==1.65.5
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.241126
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typer==0.15.2
types-pytz==2024.2.0.20241221
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xgboost==2.1.3
